PY4E: exercises
'Python for Everybody: Exploring Data Using Python 3' by Dr. Charles R. Severance.

This book teaches the essentials of Python and proposes a range of exercises along the way. This repository contains my answers. Reading this book was part of the subject ICTPRG302 - Introductory programming techniques as part of the IV Certificate in IT (Programming) at North Metropolitan TAFE, WA, Australia.